+++ descartes_light/test/utest.cpp
@@ -1,7 +1,7 @@
 #include <descartes_light/core/solver.h>
 #include <descartes_light/edge_evaluators/euclidean_distance_edge_evaluator.h>
 #include <descartes_light/solvers/ladder_graph/ladder_graph_solver.h>
-#include <descartes_light/solvers/ladder_graph/bgl_ladder_graph_solver.h>
+#include <descartes_light/solvers/bgl/bgl_ladder_graph_solver.h>
 
 #include <descartes_light/descartes_macros.h>
 DESCARTES_IGNORE_WARNINGS_PUSH
@@ -130,6 +130,7 @@ struct SolverConfigurator<BGLLadderGraphSolver<FloatT>>
   using FloatType = FloatT;
   typename Solver<FloatT>::Ptr create() { return std::make_unique<BGLLadderGraphSolver<FloatT>>(6, 1); }
 };
+
 template struct SolverConfigurator<BGLLadderGraphSolverF>;
 template struct SolverConfigurator<BGLLadderGraphSolverD>;
